put density matrix into states
